# Conflicts:
#	src/main/java/com/kh013j/CaffinoApplicationConfiguration.java
#	src/main/java/com/kh013j/controllers/login/RegisterController.java
#	src/main/java/com/kh013j/model/domain/Dish.java
#	src/main/java/com/kh013j/model/domain/Role.java
#	src/main/java/com/kh013j/model/domain/User.java
#	src/main/resources/application.properties
#	src/main/resources/database/db.changelog-master.xml
#	src/main/resources/database/test.changelog-master.xml
#	src/main/webapp/WEB-INF/jsp/Admin.jsp
#	src/main/webapp/WEB-INF/jsp/AdminReviews.jsp
#	src/main/webapp/WEB-INF/jsp/Dishdescription.jsp
#	src/main/webapp/WEB-INF/jsp/Layoutgrid.jsp
#	src/main/webapp/WEB-INF/jsp/registration.jsp
#	src/test/java/com/kh013j/DishServiceImplTest.java

+++ src/main/java/com/kh013j/model/domain/Dish.java
@@ -6,7 +6,6 @@ import javax.persistence.*;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Size;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -30,7 +29,7 @@ public class Dish {
     @Min(0)
     private int preparingtime;
     @Max(100000)
-    private BigDecimal price;
+    private int price;
     @ManyToOne
     @JoinColumn(name = "category_id")
     private Category category;
